moving files and most of references
