--- cartoframes/context/__init__.py
@@ -1,9 +1,0 @@
-from .api_context import APIContext
-
-
-def create_context(credentials):
-    return APIContext(credentials)
-
-
-__all__ = [
-]
