+++ cartoframes/data/__init__.py
@@ -1,15 +1,7 @@
-"""Data namespace contains the class to manage your data: Dataset."""
-from __future__ import absolute_import
-
-from .dataset import Dataset
-from .dataset_info import DatasetInfo
-from .registry.strategies_registry import StrategiesRegistry
-from .table import tables
-
+from .dataset.dataset import Dataset
+from .dataset.registry.strategies_registry import StrategiesRegistry
 
 __all__ = [
     'Dataset',
-    'DatasetInfo',
-    'StrategiesRegistry',
-    'tables'
+    'StrategiesRegistry'
 ]
