+++ cartoframes/data/dataset/registry/strategies_registry.py
