+++ cartoframes/data/dataset/registry/table_dataset.py
@@ -3,8 +3,8 @@ from warnings import warn
 from carto.exceptions import CartoException, CartoRateLimitException
 
 from .base_dataset import BaseDataset
-from ..columns import Column, normalize_name
-from ...utils import is_table_name
+from cartoframes.utils.columns import Column, normalize_name
+from cartoframes.utils.utils import is_table_name
 
 
 class TableDataset(BaseDataset):
