+++ cartoframes/utils/table.py
