+++ test/data/test_dataset.py
@@ -13,14 +13,14 @@ from carto.exceptions import CartoException
 from cartoframes.data import Dataset
 from cartoframes.auth import Credentials
 from cartoframes.client import SQLClient
-from cartoframes.data.utils import setting_value_exception
+from cartoframes.utils.utils import setting_value_exception
 from cartoframes.data.columns import normalize_name
 from cartoframes.utils import load_geojson
 from cartoframes.data import StrategiesRegistry
-from cartoframes.data.registry.dataframe_dataset import DataFrameDataset, _rows
-from cartoframes.data.registry.table_dataset import TableDataset
-from cartoframes.data.registry.query_dataset import QueryDataset
-from cartoframes import context
+from cartoframes.data.dataset.registry import DataFrameDataset, _rows
+from cartoframes.data.dataset.registry import TableDataset
+from cartoframes.data.dataset.registry import QueryDataset
+from cartoframes.lib import context
 
 try:
     from unittest.mock import Mock
