+++ test/mocks/dataset_mock.py
@@ -2,10 +2,10 @@
 
 import pandas as pd
 
-from cartoframes.data.registry.strategies_registry import StrategiesRegistry
-from cartoframes.data.registry.dataframe_dataset import DataFrameDataset
-from cartoframes.data.registry.query_dataset import QueryDataset
-from cartoframes.data.registry.table_dataset import TableDataset
+from cartoframes.data.dataset.registry import StrategiesRegistry
+from cartoframes.data.dataset.registry import DataFrameDataset
+from cartoframes.data.dataset.registry import QueryDataset
+from cartoframes.data.dataset.registry import TableDataset
 from cartoframes.data import Dataset, DatasetInfo
