+++ test/viz/test_map.py
@@ -2,7 +2,7 @@ import unittest
 
 from carto.exceptions import CartoException
 
-from cartoframes import context
+from cartoframes.lib import context
 from cartoframes.viz import Map, Layer, Source, constants
 from cartoframes.data import StrategiesRegistry
 from cartoframes.auth import Credentials
